Version 0.0.31 report and test extra information and not optimized for performance measure
All the benchmark datasets predicted with high G-mean, but there are couple of remaining tasks.
If the program exits normally without bugs like "Seg Fault", please read the prompt before exit.
Some of the parts which are not completely done, while exit the program with prompting to the point
which generate it.